app/requirements.txt
langgraph>=0.2.6
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-qdrant>=0.1.0
qdrant-client>=1.9.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
tiktoken>=0.7.0
pypdf>=5.0.0
pydantic>=2.0.0
